# docs/openapi-3.0.yaml
openapi: 3.0.0
info:
  contact:
    email: support@volumeviz.io
    name: API Support
    url: https://github.com/mantonx/volumeviz/issues
  description: >-
    Docker volume monitoring API with comprehensive volume discovery, size
    calculation, and container attachment tracking. Focus on user-mounted
    volumes only.


    ## Volume-First Approach

    - Automatic discovery and filtering of user-mounted volumes

    - Excludes Docker infrastructure volumes (container filesystems, tmp volumes)

    - Real-time volume usage monitoring and size calculation

    - Container attachment tracking for each volume


    ## Features

    - Multi-method volume size calculation (du, find, stat)

    - Asynchronous scanning with progress tracking for large volumes

    - Real-time WebSocket progress updates during scan operations

    - High-performance caching with TTL-based invalidation

    - Comprehensive Prometheus metrics integration

    - Circuit breaker patterns for resilience


    ## WebSocket API

    WebSocket endpoint: `ws://localhost:8080/ws`


    ### Scan Progress Events

    Subscribe to real-time scan progress updates:

    ```json

    {

    "type": "subscribe",

    "event": "scan_progress",

    "data": {

    "volume_id": "volume_id_here",

    "scan_id": "scan_id_here"

    }

    }

    ```


    Progress updates are broadcast as:

    ```json

    {

    "type": "scan_progress_update",

    "data": {

    "scan_id": "scan_id",

...

...

...

    "phases": [

    {

...

...

...
paths:
...
  /api/v1/search/saved:
    get:
      description: Get a list of all saved searches
      parameters:
        - description: Page number
          in: query
          name: page
          schema:
            type: integer
        - description: Items per page
          in: query
          name: perPage
          schema:
            type: integer
        - description: Filter by tags
          in: query
          name: tags
          style: form
          explode: false
          schema:
            type: array
            items:
              type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_api_v1_search.ListSavedSearchesResponse"
      summary: List saved searches
      tags:
        - Search
    post:
      description: Save a search query for later use
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/internal_api_v1_search.CreateSavedSearchRequest"
        description: Saved search details
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_api_v1_search.SavedSearch"
      summary: Create a saved search
      tags:
        - Search
  "/api/v1/search/saved/{id}":
    delete:
      description: Delete a saved search by ID
      parameters:
        - description: Saved search ID
          in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content
      summary: Delete a saved search
      tags:
        - Search
    get:
      description: Get details of a specific saved search
      parameters:
        - description: Saved search ID
          in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_api_v1_search.SavedSearch"
      summary: Get a saved search
      tags:
        - Search
    put:
      description: Update an existing saved search
      parameters:
        - description: Saved search ID
          in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/internal_api_v1_search.UpdateSavedSearchRequest"
        description: Updated search details
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_api_v1_search.SavedSearch"
      summary: Update a saved search
      tags:
        - Search
  "/api/v1/search/saved/{id}/run":
    post:
      description: Execute a saved search and return results
      parameters:
        - description: Saved search ID
          in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_api_v1_search.SearchFilesResponse"
      summary: Run a saved search
      tags:
        - Search
...
servers:
...
components:
  schemas:
...
    internal_api_v1_search.CreateSavedSearchRequest:
      properties:
        description:
          type: string
        is_public:
          type: boolean
        metadata:
          additionalProperties: true
          type: object
        name:
          type: string
        query:
          $ref: "#/components/schemas/internal_api_v1_search.SearchFilesRequest"
        tags:
          items:
            type: string
          type: array
      required:
        - name
        - query
      type: object
    internal_api_v1_search.FileResult:
      properties:
        audio_codec:
          type: string
        camera_model:
          type: string
        capture_date:
          type: string
        created_time:
          type: string
        disk_usage:
          type: integer
        duration_ms:
          description: Media metadata
          type: integer
        extension:
          type: string
        gps_lat:
          type: number
        gps_lon:
          type: number
        has_gps:
          type: boolean
        height:
          type: integer
        id:
          type: integer
        media_kind:
          type: string
        metadata:
          additionalProperties: true
          type: object
        mime_type:
          type: string
        modified_time:
          type: string
        name:
          type: string
        path:
          type: string
        preview_url:
          type: string
        size:
          type: integer
        video_codec:
          type: string
        volume_id:
          type: string
        width:
          type: integer
      type: object
    internal_api_v1_search.ListSavedSearchesResponse:
      properties:
        page:
          type: integer
        per_page:
          type: integer
        searches:
          items:
            $ref: "#/components/schemas/internal_api_v1_search.SavedSearch"
          type: array
        total_count:
          type: integer
      type: object
    internal_api_v1_search.SavedSearch:
      properties:
        created_at:
          type: string
        description:
          type: string
        id:
          type: integer
        is_public:
          type: boolean
        last_run_at:
          type: string
        metadata:
          additionalProperties: true
          type: object
        name:
          type: string
        query:
          $ref: "#/components/schemas/internal_api_v1_search.SearchFilesRequest"
        run_count:
          type: integer
        tags:
          items:
            type: string
          type: array
        updated_at:
          type: string
      required:
        - name
      type: object
    internal_api_v1_search.SearchFilesRequest:
      properties:
        durationFrom:
          description: Media metadata filters
          type: integer
        durationTo:
          description: Max duration in ms
          type: integer
        glob:
          description: Glob pattern for path matching
          type: string
        hasGps:
          description: Has GPS coordinates
          type: boolean
        hasSubs:
          description: Has subtitles
          type: boolean
        hashPresent:
          description: Has computed hash
          type: boolean
        maxHeight:
          description: Max height in pixels
          type: integer
        maxSize:
          description: Maximum file size in bytes
          type: integer
        maxWidth:
          description: Max width in pixels
          type: integer
        mediaKind:
          description: Media filters
          type: string
        mime:
          description: MIME type filters
          items:
            type: string
          type: array
        minHeight:
          description: Min height in pixels
          type: integer
        minSize:
          description: Size filters
          type: integer
        minWidth:
          description: Min width in pixels (video/image)
          type: integer
        mtimeFrom:
          description: Time filters
          type: string
        mtimeTo:
          description: Modified time to
          type: string
        order:
          description: "Sort order: asc, desc"
          type: string
        page:
          description: Pagination and sorting
          minimum: 1
          type: integer
        path:
          description: Path prefix filter
          type: string
        perPage:
          maximum: 100
          minimum: 1
          type: integer
        q:
          description: Text search
          type: string
        regex:
          description: Regex pattern for path matching
          type: string
        sort:
          description: "Sort field: relevance, name, size, mtime, ctime, duration, type,
            media_kind"
          type: string
      type: object
    internal_api_v1_search.SearchFilesResponse:
      properties:
        files:
          items:
            $ref: "#/components/schemas/internal_api_v1_search.FileResult"
          type: array
        filters: {}
        page:
          type: integer
        per_page:
          type: integer
        query_time_ms:
          type: integer
        total_count:
          type: integer
        total_pages:
          type: integer
      type: object
...
    internal_api_v1_search.UpdateSavedSearchRequest:
      properties:
        description:
          type: string
        is_public:
          type: boolean
        metadata:
          additionalProperties: true
          type: object
        name:
          type: string
        query:
          $ref: "#/components/schemas/internal_api_v1_search.SearchFilesRequest"
        tags:
          items:
            type: string
          type: array
      type: object
...